requirements.txt
beautifulsoup4==4.11.1
camelot-py==0.10.1
pandas==1.5.0
python-telegram-bot==13.14